Refactor: Move helper functions and components to separate files
